requirements.txt
Django==1.7
argparse==1.2.1
dj-database-url==0.3.0
gevent==1.0.1
greenlet==0.4.4
gunicorn==19.1.1
httplib2==0.9
psycopg2==2.5.4
python-instagram==1.1.3
setproctitle==1.1.8
simplejson==3.6.4
wsgiref==0.1.2
